rrr: convert rpc/tcp_channel to module (rrr.tcp_channel)
Merges tcp_channel.hpp (528 lines) + tcp_channel.cpp (946 lines) into
the single module unit src/rrr/rpc/tcp_channel.cpp (~1430 lines).
Exports the `TcpConnection` / `TcpListener` / `TcpFactory` classes and
their channel / pollable adapter glue. Sole textual consumer (rrr.hpp
umbrella) now imports rrr.tcp_channel instead of including the .hpp.

Note: clang 22 happens to be happy with this module unit (no
`Reactor::create_sp_event<>()` calls in its purview). The sibling
fiber_channel conversion remains deferred because of a clang 22
codegen crash on cross-module template-method instantiation in
module purview — see plan doc and the open task.
